scalar_type set to long double..

+++ src/ALE.h
@@ -28,7 +28,7 @@
 
 typedef bpp::TreeTemplate<bpp::Node> tree_type;
 //typedef long double scalar_type;
-typedef double scalar_type;
+typedef  long double scalar_type;
 typedef std::pair<bpp::Node*, bpp::Node*> dedge_type;
 //typedef std::pair<long int,std::set < long int > > dip_type;
